different return type

+++ speck-prng.h
@@ -3,6 +3,6 @@
 
 void generate_block_ctr(size_t length, uint8_t *block, uint8_t *seed);
 
-uint8_t *get_seed_64();
+uint64_t * get_seed_64();
 
 #endif
